version 2
I've changed the layout of the quiz to only show one question at a time as suggested. The quiz seems to function well, but if you see any bugs that I've missed it would be useful! I haven't managed to incorperate every suggestion just yet. But maybe now would be a good time for another assesment for further improvements.

I'll hold off on looking at your quiz for now Dan. Maybe it would be best for you to take another look at this version before I do, so as not to get all the answers before trying to figure it out first.

Anyway, whenever you get a chance mate, thanks again for all the help!